% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSimulations.R
\name{ApplyAcross_m_and_p}
\alias{ApplyAcross_m_and_p}
\title{Sim across m and p}
\usage{
ApplyAcross_m_and_p(Sim, ntrain, ntest, pvec, mvec, contamination = "Var",
  Vartype = "Id", DGP, ntrees, ndsize, ntreestune, parvec, cvreps, cvfolds,
  tol)
}
\arguments{
\item{Sim}{Which simulation? Either "RL" for Roy Larocque (2012), or "LM" for Li, Martin (2017)}

\item{ntrain}{number of training cases}

\item{ntest, }{number of test cases}

\item{pvec}{proportion of outliers}

\item{mvec}{value of m to use if Sim == "RL"}

\item{contamination}{Use either variance ("Var") or mean ("Mean") contamination. Only relevant for Sim =="RL".}

\item{Vartype}{use identity ("Id") or Toeplitz ("Toeplitz") correlation matrix. Only relevant for Sim =="LM"}

\item{DGP}{If Sim == "RL", which data generating process should be used? either 1 for tree-like, or 2 for non-tree}

\item{ntrees}{number of trees}

\item{ndsize}{nodesize}

\item{ntreestune}{number of trees to use for tuning alpha}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{cvfolds}{number of repetitions to perform in cross validation}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}

\item{DATA}{object from generate_RLdata or generate_LMdata}

\item{cvfolds}{number of folds to perform in cross validation}
}
\value{
returns a list of 4 items
       1. Datasets (TRAIN, TEST, and Outlier Indicator)
       2. Matrix of 16 columns giving different predictions. Last column is true Y.
       3. Number of iterations
       4. Output from TuneMultifoldCV (a list of 8 items itself)
}
\description{
Function to repeat simulation over vectors for m and p
}
