% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealData.R
\name{Assess_Real_Data}
\alias{Assess_Real_Data}
\title{Assess_Real_Data}
\usage{
Assess_Real_Data(dataset, nfolds, p, ntrees, ndsize, ntreestune, parvec, cvreps,
  cvfolds, tol)
}
\arguments{
\item{dataset}{real dataset to work with}

\item{nfolds}{number of folds to partition into to assess performance}

\item{p}{percentage of training cases for which to add contamination (using N(0, 5*sd(Y)))}

\item{ntrees}{number of trees}

\item{ndsize}{nodesize}

\item{ntreestune}{number of trees to use for tuning alpha}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{cvreps}{number of repetitions to perform in cross validation}

\item{cvfolds}{number of folds to perform in cross validation}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}
}
\value{
returns a list of 4 items
       1. Datasets (TRAIN, TEST, and Outlier Indicator)
       2. Matrix of 16 columns giving different predictions. Last column is true Y.
       3. Number of iterations
       4. Output from TuneMultifoldCV (a list of 8 items itself)
}
\description{
Function to divide real datasets into test and training sets, possibly add contamination
and assess predictions
}
