% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{AssignFolds}
\alias{AssignFolds}
\title{Partition training data}
\usage{
AssignFolds(ntrain, cvfolds)
}
\arguments{
\item{ntrain}{number of training cases}

\item{cvfolds}{number of folds for cross validation (must evenly divide number of training cases)}
}
\value{
returns an array containing
1- indices of training cases
2- fold numbers
}
\description{
Function to partition training cases into folds for a given number of repetitions of cross validation
Will typically be called within PartitionData, which is used in parameter tuning step.
}
