% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Calc_fold_CV_Error}
\alias{Calc_fold_CV_Error}
\title{Apply Parameter Tuning}
\usage{
Calc_fold_CV_Error(TRAIN, TestInd, fold, ndsize, OutlierInd, parvec,
  tol = 10^-4, ntreestune = ntreestune)
}
\arguments{
\item{TRAIN}{matrix of training cases with response in last column}

\item{TestInd}{Matrix indicating which cases are in test (validation) set, TRAIN2}

\item{fold}{number of fold within cross validation}

\item{ndsize}{nodesize random forest tuning parameter for cross validation}

\item{OutlierInd}{Vector of zeros and ones indicating whether training cases came from contaminating distribution}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}

\item{ntreestune}{number of trees to use for forests involved in parameter tuning}
}
\value{
Returns array of dimensions
6 by 3 by length(parvec) containing errors. Dimensions index (1) type of error calculated
         1-MSE without downweighting outliers in CV error
         2-MAPE without downweighting outliers in CV error
         3-MSE downweighting outliers according to BisqwtRF
         4-MAPE downweighting outliers according to BisqwtRF
         5-MSE downweighting outliers according to BisqwtRFL
         6-MAPE downweighting outliers according to BisqwtRFL
(2) applied to all cases in TRAIN2, outliers only, nonoutliers only
(3) index of alpha from parvec
}
\description{
For each fold in the cross validation, calls Get_CVWeights to split into training and validation data
and get weights, then applies Calculate_CV_Error over all candidate values for alpha. Returns array of dimensions
6 by 3 by length(parvec) containing errors. Dimensions index (1) type of error calculated
         1-MSE without downweighting outliers in CV error
         2-MAPE without downweighting outliers in CV error
         3-MSE downweighting outliers according to BisqwtRF
         4-MAPE downweighting outliers according to BisqwtRF
         5-MSE downweighting outliers according to BisqwtRFL
         6-MAPE downweighting outliers according to BisqwtRFL
(2) applied to all cases in TRAIN2, outliers only, nonoutliers only
(3) index of alpha from parvec
}
