% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Calculate_CV_Error}
\alias{Calculate_CV_Error}
\title{Parameter tuning function
#'
For each candidate parameter value, calculate the CV error resulting from growing a forest on TRAIN1,
predicting TRAIN2, and downweighting contributions in TRAIN2 based on either RF residuals or RFL residuals
using OOB predictions when TRAIN2 was predicted using RF grown on TRAIN2. In cross validation, this is
done for each candidate value, within each fold, within each rep.}
\usage{
Calculate_CV_Error(OOBWeights1, PredWeights1, TRAINY, samp, parvec, ind,
  tol = 10^-4, BisqwtRF, BisqwtRFL, OutlierInd)
}
\arguments{
\item{OOBWeights1}{OOB prediction weights for training cases in TRAIN1}

\item{PredWeights1}{Prediction weights for predicting test (validation) cases using training cases}

\item{TRAINY}{responses for all training cases}

\item{samp}{list of indexes of test (or validation) cases (i.e. TRAIN2)}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{ind}{index of parameter tuning vector to work on}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}

\item{BisqwtRF}{Weight to be applied to each test or validation case when using cross validation to set tuning parameter, using RF outliers for downweighting}

\item{OutlierInd}{Vector of zeros and ones indicating whether training cases came from contaminating distribution}
}
\value{
Returns a 6 by 3 matrix with errors. Rows indicate type of weighting applied to errors in TRAIN2.
         1-MSE without downweighting outliers in CV error
         2-MAPE without downweighting outliers in CV error
         3-MSE downweighting outliers according to BisqwtRF
         4-MAPE downweighting outliers according to BisqwtRF
         5-MSE downweighting outliers according to BisqwtRFL
         6-MAPE downweighting outliers according to BisqwtRFL
        Columns represent error for all cases (1), only outliers (2), only nonoutliers(3)
}
\description{
Parameter tuning function
#'
For each candidate parameter value, calculate the CV error resulting from growing a forest on TRAIN1,
predicting TRAIN2, and downweighting contributions in TRAIN2 based on either RF residuals or RFL residuals
using OOB predictions when TRAIN2 was predicted using RF grown on TRAIN2. In cross validation, this is
done for each candidate value, within each fold, within each rep.
}
