% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustPreds.R
\name{Compute_deltas}
\alias{Compute_deltas}
\title{Function to compute RFLOWESS weight multipliers for training cases}
\usage{
Compute_deltas(OOBWeights, TRAINY, alpha = 6, method = "Tukey",
  tol = 10^-6, maxiter = 1000)
}
\arguments{
\item{OOBWeights}{matrix of OOB training case weights}

\item{TRAINY}{vector of training case response values}

\item{alpha}{tuning parameter for robust weighting. Small alpha corresponds to more aggressive robust adjustment.}

\item{method}{should Tukey or Huber weighting function be used?}

\item{tol}{maximal tolerance for change between iterations}

\item{maxiter}{maximum number of iterations}
}
\value{
returns a list containing
1- vector of multipliers for training case weights
2- adjusted OOB training case predictions
3- number of iterations
}
\description{
Iteratively adjusts training case weights using residuals for out-of-bag cases.
}
