% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Evaluate_Tuning_Candidates}
\alias{Evaluate_Tuning_Candidates}
\title{Evaluate Tuning Candidates}
\usage{
Evaluate_Tuning_Candidates(TRAIN, OutlierInd, cvfolds, parvec, ndsize,
  ntreestune = 100, tol = 10^-4)
}
\arguments{
\item{TRAIN}{matrix of training cases with response in last column}

\item{OutlierInd}{Vector of zeros and ones indicating whether training cases came from contaminating distribution}

\item{cvfolds}{number of folds to perform in cross validation}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{ndsize}{nodesize random forest tuning parameter for cross validation}

\item{ntreestune}{number of trees to use for forests involved in parameter tuning}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}
}
\value{
Returns array of dimensions
6 by 3 by length(parvec) by number of folds, containing errors. Dimensions index (1) type of error calculated
         1-MSE without downweighting outliers in CV error
         2-MAPE without downweighting outliers in CV error
         3-MSE downweighting outliers according to BisqwtRF
         4-MAPE downweighting outliers according to BisqwtRF
         5-MSE downweighting outliers according to BisqwtRFL
         6-MAPE downweighting outliers according to BisqwtRFL
(2) applied to all cases in TRAIN2, outliers only, nonoutliers only
(3) index of alpha from parvec
(4) index of fold in cross validation
}
\description{
This function applies Calc_fold_CV_Error over all folds of a cross validation. Use replicate if running multiple
repetitions of cross-validation
}
