% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{generate_LMdata}
\alias{generate_LMdata}
\title{Generate Li, Martin data}
\usage{
generate_LMdata(ntrain, ntest, p, Vartype = "Id")
}
\arguments{
\item{ntrain}{number of training cases to generate}

\item{ntest}{number of test cases to generate}

\item{p}{probability of coming from contaminating distribution}

\item{Vartype}{covariance matrix can be either identity matrix ("Id") or Toeplitz(0.7) matrix ("Toeplitz)}
}
\value{
returns a list containing the training dataset, the test dataset, and an indicator of whether training cases came from contaminating distribution.
}
\description{
Function to generate data from simulations used in Li & Martin (2017)
}
