% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{generate_RLdata}
\alias{generate_RLdata}
\title{Generate Roy, Larocque data}
\usage{
generate_RLdata(ntrain, ntest, p, m, type = "Var", DGP = 1)
}
\arguments{
\item{ntrain}{number of training cases to generate}

\item{ntest}{number of test cases to generate}

\item{p}{probability of coming from contaminating distribution}

\item{m}{multiplier determining signal-to-noise ratio}

\item{type}{should contamination be done to the variance or the mean}

\item{DGP}{Which data generating process from Roy & Larocque (2012) should be used? Choices are 1 or 2.}
}
\value{
returns a list containing the training dataset, the test dataset, and an indicator of whether training cases came from contaminating distribution.
}
\description{
Function to generate data from simulations used in Roy & Larocque (2012)
}
