% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustPreds.R
\name{Huber}
\alias{Huber}
\title{Pseudo-Huber function}
\usage{
Huber(t)
}
\arguments{
\item{t}{& a vector of values on which to apply the pseudo Huber function}
}
\value{
returns a vector the same length as t, containing values after applying pesudo Huber function
}
\description{
Pseudo-Huber function, used by Li & Martin, applied to a vector of values, element-by-element
}
