% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustPreds.R
\name{LiPred}
\alias{LiPred}
\title{Li, Martin Predictions}
\usage{
LiPred(OOBWeights, PredWeights, TRAINY, method = "Huber", delta = 0.005,
  tol = 10^-6, maxiter = 100)
}
\arguments{
\item{OOBWeights}{matrix of OOB training case weights}

\item{PredWeights}{ntest by ntrain matrix of prediction case weights}

\item{TRAINY}{vector of training case response values}

\item{method}{should Tukey or Huber weighting function be used?}

\item{delta}{value of tuning parameter}

\item{tol}{maximal tolerance for change between iterations}

\item{maxiter}{maximum number of iterations}
}
\value{
returns a list containing
1- vector of predictions
2- vector of adjusted case weights
3- number of iterations
}
\description{
Predictions obtained by reweighting training cases using Li & Martin's (2017) approach
}
