% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustPreds.R
\name{LOWESSPred}
\alias{LOWESSPred}
\title{RFLOWESS Prediction}
\usage{
LOWESSPred(OOBWeights, PredWeights, TRAINY, alpha = 6, method = "Tukey",
  tol = 10^-6)
}
\arguments{
\item{OOBWeights}{matrix of OOB training case weights}

\item{PredWeights}{ntest by ntrain matrix of prediction case weights}

\item{TRAINY}{vector of training case response values}

\item{method}{should Tukey or Huber weighting function be used?}

\item{tol}{maximal tolerance for change between iterations}

\item{maxiter}{maximum number of iterations}
}
\value{
returns a list containing
1- vector of LOWESSRF predictions
2- vector of adjusted case weights
3- number of iterations
}
\description{
Function to obtain LOWESSRF prediction using case weights for OOB predictions and test predictions
by calling Compute_deltas
}
