% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSimulations.R
\name{Make_All_Preds}
\alias{Make_All_Preds}
\title{Make Predictions}
\usage{
Make_All_Preds(DATA, ntrees, ndsize, ntreestune, parvec, cvreps, cvfolds, tol)
}
\arguments{
\item{DATA}{object from generate_RLdata or generate_LMdata}

\item{ntrees}{number of trees}

\item{ndsize}{nodesize}

\item{ntreestune}{number of trees to use for tuning alpha}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{cvfolds}{number of repetitions to perform in cross validation}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}

\item{cvfolds}{number of folds to perform in cross validation}
}
\value{
returns a list of 4 items
       1. Datasets (TRAIN, TEST, and Outlier Indicator)
       2. Matrix of 16 columns giving different predictions. Last column is true Y.
       3. Number of iterations
       4. Output from TuneMultifoldCV (a list of 8 items itself)
}
\description{
Function to make predictions for all methods tested in simulation
}
