% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{MakeLowessPred}
\alias{MakeLowessPred}
\title{Predict RFLOWESS Tuning}
\usage{
MakeLowessPred(OOBWeights, PredWeights, TRAINY, TEST, tol = tol, ind, parvec,
  method = "Tukey")
}
\arguments{
\item{OOBWeights}{matrix of training cases with response in last column}

\item{PredWeights}{Vector of zeros and ones indicating whether training cases came from contaminating distribution}

\item{TRAINY}{number of folds to perform in cross validation}

\item{TEST}{Test data}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}

\item{ind}{index of parameter vector to use in tuning}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{method}{should Tukey or Huber weighting function be used?}

\item{ndsize}{nodesize random forest tuning parameter for cross validation}
}
\value{
Returns:
        LPREDERR: A length 2 vector containing MSPE and MAPE on test data using alpha parameter specified
        LWeights: A ntest by ntrain matrix containing LOWESSRF weights for test cases
        LIter: Number of iterations until convergence of LOWESSRF algorithm
}
\description{
This is a function to be called inside TuneMultifoldCV. It makes the RFLOWESS predictions for a specific alpha
and records them in a matrix LPREDERR.
}
