% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustPreds.R
\name{Node_Tree_Agg}
\alias{Node_Tree_Agg}
\title{Median Aggregation Predictions.}
\usage{
Node_Tree_Agg(TrainNodes, TestNodes, Inbag, TRAINY)
}
\arguments{
\item{TrainNodes}{ntrain by ntree matrix containing terminal node numbers of each training case in each tree}

\item{TestNodes}{ntest by ntree matrix containing terminal node numbers of each training case in each tree}

\item{Inbag}{ntrain by ntree matrix telling number of times training case came up in bootstrap sample used to grow each tree}

\item{TRAINY}{vector of training case response values}
}
\value{
returns a list containing predictions using
1- mean within terminal nodes and median across trees
2- median within terminal nodes and median across trees
3- median within terminal nodes and mean across trees
}
\description{
Method to get predictions from median-type aggregation methods used in paper by Roy & Larocque (2012)
}
