% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{Partition_Real_Data}
\alias{Partition_Real_Data}
\title{Partition Real Data}
\usage{
Partition_Real_Data(dataset, nfolds)
}
\arguments{
\item{dataset}{real dataset to work with}

\item{nfolds}{number of folds to partition into to assess performance}
}
\value{
returns a list containing the training dataset, the test dataset, and an indicator of whether training cases came from contaminating distribution.
}
\description{
Function to partition real datasets into test and training sets and possibly add contamination
}
