% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSimulations.R
\name{RunSimulation}
\alias{RunSimulation}
\title{Run Simulation}
\usage{
RunSimulation(Sim = "RL", ntrain, ntest, p, m, contamination = "Var",
  Vartype = "Id", DGP = 2, ntrees, ndsize, ntreestune, parvec, cvreps,
  cvfolds, tol)
}
\arguments{
\item{Sim}{Which simulation? Either "RL" for Roy Larocque (2012), or "LM" for Li, Martin (2017)}

\item{ntrain}{number of training cases}

\item{ntest, }{number of test cases}

\item{p}{proportion of outliers}

\item{m}{signal to noise parameter when Sim=="RL"}

\item{contamination}{Use either variance ("Var") or mean ("Mean") contamination. Only relevant for Sim =="RL".}

\item{Vartype}{use identity ("Id") or Toeplitz ("Toeplitz") correlation matrix. Only relevant for Sim =="LM"}

\item{DGP}{If Sim == "RL", which data generating process should be used? either 1 for tree-like, or 2 for non-tree}

\item{ntrees}{number of trees}

\item{ndsize}{nodesize}

\item{ntreestune}{number of trees to use for tuning alpha}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{cvreps}{number of repetitions to perform in cross validation}

\item{cvfolds}{number of folds to perform in cross validation}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}

\item{DATA}{object from generate_RLdata or generate_LMdata}
}
\value{
returns a list of 4 items
       1. Datasets (TRAIN, TEST, and Outlier Indicator)
       2. Matrix of 16 columns giving different predictions. Last column is true Y.
       3. Number of iterations
       4. Output from TuneMultifoldCV (a list of 8 items itself)
}
\description{
Function to generate data and run full simulaton
}
