% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustPreds.R
\name{Tukey}
\alias{Tukey}
\title{Tukey bisquare function}
\usage{
Tukey(t)
}
\arguments{
\item{t}{& a vector of values on which to apply the Tukey Bisquare function}
}
\value{
returns a vector the same length as t, containing values after applying Tukey bisquare function
}
\description{
Bisquare function applied to a vector of values, element-by-element
}
