% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealData.R
\name{Assess_Folds}
\alias{Assess_Folds}
\title{Assess_Folds}
\usage{
Assess_Folds(dataset, partitions, p, fold, ntrees, ndsize, ntreestune, parvec,
  cvreps, cvfolds, tol)
}
\arguments{
\item{dataset}{List of partitioned dataframes for training, test sets, and indicator of outlying observations}

\item{partitions}{array containing indices of training and test cases}

\item{p}{percentage of training cases for which to add contamination (using N(0, 5*sd(Y)))}

\item{fold}{fold on which to assess performance}

\item{ntrees}{number of trees}

\item{ndsize}{nodesize}

\item{ntreestune}{number of trees to use for tuning alpha}

\item{parvec}{vector of candidate values for tuning parameter alpha}

\item{cvreps}{number of repetitions to perform in cross validation}

\item{cvfolds}{number of folds to perform in cross validation}

\item{tol}{maximal change in interation for LOWESSRF weights in cross validation}
}
\value{
returns a list of 4 items
       1. Datasets (TRAIN, TEST, and Outlier Indicator)
       2. Matrix of 16 columns giving different predictions. Last column is true Y.
       3. Number of iterations
       4. Output from TuneMultifoldCV (a list of 8 items itself)
}
\description{
Function to take partitioned data, and make predictions across all folds
}
